+++ src/views/autoCompleteView.js
@@ -1,4 +1,10 @@
 let resultsList;
+const dataAttribute = "data-result";
+const select = {
+	resultsList: "autoComplete_results_list",
+	result: "autoComplete_result",
+	highlight: "autoComplete_highlighted"
+};
 
 // Gets the user's input value
 const getInput = selector => (typeof selector === "string" ? document.querySelector(selector) : selector());
@@ -6,23 +12,51 @@ const getInput = selector => (typeof selector === "string" ? document.querySelec
 // Creats the results list HTML tag
 const createResultsList = renderResults => {
 	resultsList = document.createElement("ul");
-	resultsList.setAttribute("id", "autoComplete_results_list");
+	resultsList.setAttribute("id", select.resultsList);
 	renderResults.destination.insertAdjacentElement(renderResults.position, resultsList);
 };
 
 // Hightlight matching values
-const highlight = value => `<span class="autoComplete_highlighted">${value}</span>`;
+const highlight = value => `<span class=${select.highlight}>${value}</span>`;
 
 // Adding matching results to the list
 const addResultsToList = (dataSrc, dataKey) => {
 	dataSrc.forEach((event, record) => {
 		const result = document.createElement("li");
 		const resultValue = dataSrc[record].source[dataKey] || dataSrc[record].source;
-		result.setAttribute("autoComplete-data", resultValue);
-		result.setAttribute("class", "autoComplete_result");
+		result.setAttribute(dataAttribute, resultValue);
+		result.setAttribute("class", select.result);
+		result.setAttribute("tabindex", "1");
 		result.innerHTML = dataSrc[record].match || dataSrc[record];
 		resultsList.appendChild(result);
 	});
+};
+
+// Keyboard Arrow Navigation
+const navigation = selector => {
+	const input = getInput(selector);
+	const first = resultsList.firstChild;
+	document.onkeydown = event => {
+		const active = document.activeElement;
+		switch (event.keyCode) {
+		// Arrow Up
+		case 38:
+			if (active !== first && active !== input) {
+				active.previousSibling.focus();
+			} else if (active === first) {
+				input.focus();
+			}
+			break;
+			// Arrow Down
+		case 40:
+			if (active === input && resultsList.childNodes.length > 0) {
+				first.focus();
+			} else if (active !== resultsList.lastChild) {
+				active.nextSibling.focus();
+			}
+			break;
+		}
+	};
 };
 
 // Clears the list of results
@@ -30,21 +64,28 @@ const clearResults = () => (resultsList.innerHTML = "");
 
 // Gets user selection
 const getSelection = (field, callback, resultsValues, dataKey) => {
-	const results = resultsList.querySelectorAll(".autoComplete_result");
+	const results = resultsList.querySelectorAll(`.${select.result}`);
 	results.forEach(selection => {
-		selection.addEventListener("mousedown", event => {
-			// Callback function invoked on user selection
-			callback({
-				query: getInput(field).value,
-				results: resultsValues.map(record => record.source),
-				selection: resultsValues.find(value => {
-					const resValue = value.source[dataKey] || value.source;
-					return resValue === event.target.closest(".autoComplete_result")
-						.getAttribute("autoComplete-data");
-				}).source
+		["mousedown", "keydown"].forEach(eventType => {
+			selection.addEventListener(eventType, event => {
+				if (eventType === "mousedown" || event.keyCode === 13) {
+					// Callback function invoked on user selection
+					callback({
+						event,
+						query: getInput(field).value,
+						results: resultsValues.map(record => record.source),
+						selection: resultsValues.find(value => {
+							const resValue = value.source[dataKey] || value.source;
+							return (
+								resValue === event.target.closest(`.${select.result}`)
+									.getAttribute(dataAttribute)
+							);
+						}).source
+					});
+					// Clear Results after selection is made
+					clearResults();
+				}
 			});
-			// Clear Results after selection is made
-			clearResults();
 		});
 	});
 };
@@ -54,6 +95,7 @@ export const autoCompleteView = {
 	createResultsList,
 	highlight,
 	addResultsToList,
+	navigation,
+	clearResults,
 	getSelection,
-	clearResults
 };